p-bot/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
agno>=1.3.0
python-multipart>=0.0.5
pillow>=9.0.0
httpx>=0.24.0
pyyaml>=6.0
typing-extensions>=4.5.0
openai>=1.10.0
anthropic>=0.8.0
duckduckgo-search>=4.5
lancedb>=0.6.0
pandas==2.2.3
pypdf>=4.1.0
aiofiles>=23.2.1
numpy>=1.24.0 